+++ packages/config/src/configs/bifrostKusama.ts
@@ -4,7 +4,7 @@ import {
   ExtrinsicBuilder,
   FeeBuilder,
 } from '@moonbeam-network/xcm-builder';
-import { bnc, movr } from '../assets';
+import { bnc, movr, vbnc, vksm, vmovr } from '../assets';
 import { bifrostKusama, moonriver } from '../chains';
 import { AssetConfig } from '../types/AssetConfig';
 import { ChainConfig } from '../types/ChainConfig';
@@ -32,6 +32,54 @@ export const bifrostKusamaConfig = new ChainConfig({
         balance: BalanceBuilder().substrate().tokens().accounts(),
       },
       extrinsic: ExtrinsicBuilder().xTokens().transfer(),
+      fee: {
+        asset: bnc,
+        balance: BalanceBuilder().substrate().system().account(),
+      },
+      min: AssetMinBuilder().assetRegistry().currencyMetadatas(),
+    }),
+    new AssetConfig({
+      asset: vbnc,
+      balance: BalanceBuilder().substrate().tokens().accounts(),
+      destination: moonriver,
+      destinationFee: {
+        amount: 0.2,
+        asset: bnc,
+        balance: BalanceBuilder().substrate().system().account(),
+      },
+      extrinsic: ExtrinsicBuilder().xTokens().transferMultiCurrencies(),
+      fee: {
+        asset: bnc,
+        balance: BalanceBuilder().substrate().system().account(),
+      },
+      min: AssetMinBuilder().assetRegistry().currencyMetadatas(),
+    }),
+    new AssetConfig({
+      asset: vksm,
+      balance: BalanceBuilder().substrate().tokens().accounts(),
+      destination: moonriver,
+      destinationFee: {
+        amount: 0.2,
+        asset: bnc,
+        balance: BalanceBuilder().substrate().system().account(),
+      },
+      extrinsic: ExtrinsicBuilder().xTokens().transferMultiCurrencies(),
+      fee: {
+        asset: bnc,
+        balance: BalanceBuilder().substrate().system().account(),
+      },
+      min: AssetMinBuilder().assetRegistry().currencyMetadatas(),
+    }),
+    new AssetConfig({
+      asset: vmovr,
+      balance: BalanceBuilder().substrate().tokens().accounts(),
+      destination: moonriver,
+      destinationFee: {
+        amount: 0.2,
+        asset: bnc,
+        balance: BalanceBuilder().substrate().system().account(),
+      },
+      extrinsic: ExtrinsicBuilder().xTokens().transferMultiCurrencies(),
       fee: {
         asset: bnc,
         balance: BalanceBuilder().substrate().system().account(),